+++ base_aux/base_nest_dunders/m1_init3_params_dict_kwargs_update.py
@@ -1,3 +1,6 @@
+from typing import *
+import pytest
+
 from base_aux.aux_dict.m2_dict_ic import *
 
 
@@ -20,15 +23,18 @@ class NestInit_ParamsDict_UpdateKwargs:
 
 
 # =====================================================================================================================
-if __name__ == "__main__":
-    pass
-
+def test__():
     class Victim(NestInit_ParamsDict_UpdateKwargs):
         PARAMS = DictIc_LockedKeys(attr1=1, attr2=2)
 
     victim = Victim(attr2=22)
     assert victim.PARAMS["attr1"] == 1
     assert victim.PARAMS["attr2"] == 22
+
+
+# =====================================================================================================================
+if __name__ == "__main__":
+    pass
 
 
 # =====================================================================================================================