+++ HISTORY.md
@@ -358,7 +358,7 @@
 
 0.0.15 (2024/11/12 18:04:29)
 ------------------------------
-- [DictGaAnnotRequired] add  
+- [DictIcKeys_Ga_AnnotRequired] add  
 - [_SAND] create and add first content  
 
 0.0.14 (2024/11/08 17:51:33)
@@ -398,7 +398,7 @@
 
 0.0.5 (2024/11/05 10:33:43)
 ------------------------------
-- [classes] add DictIcKeys+DictGa  
+- [classes] add DictIcKeys+DictIcKeys_Ga  
 - [stock] move here  
 - [all] fix almost, so tests works (need fix privates for alerts)